+++ libraries/template-haskell/changelog.md
@@ -24,6 +24,8 @@
 
   * Add `Semigroup` and `Monoid` instances for `Q` (#18123).
 
+  * Add `MonadFix` instance for `Q` (#12073).
+
 ## 2.16.0.0 *TBA*
 
   * Add support for tuple sections. (#15843) The type signatures of `TupE` and